Remake from structure to class.
