+++ CMakeLists.txt
@@ -55,6 +55,7 @@ set(GITREV_TMP ${CMAKE_BINARY_DIR}/${GITREV_BARE_TMP})
 add_executable(
 	Prusa-CW-Firmware
 	Prusa-CW-Firmware/Countimer.cpp
+	Prusa-CW-Firmware/Selftest.cpp
 	Prusa-CW-Firmware/main.cpp
 	Prusa-CW-Firmware/MCP23S17.cpp
 	Prusa-CW-Firmware/thermistor.cpp
